Create Standalone MockCovidAPIImpl
